Initial inclusion of ctemplate
Summary:
Initial inclusion of the ctemplate C++ library. This is required for Kraft.
I am using the latest git commit for python3 support. A new release has not yet been for python3, but this will use that tag when it happens.

Ref T8617

Test Plan: Build and test Kraft using this package.

Reviewers: #triage_team, JoshStrobl

Reviewed By: #triage_team, JoshStrobl

Subscribers: JoshStrobl

Maniphest Tasks: T8617

Differential Revision: https://dev.getsol.us/D8113